Detection/requirements.txt
opencv-python
pycocotools
torchvision